make CI green: git deps for arb-alloy/arb-revm, harden CI, fix warnings
- flip sibling path deps to git deps pinned at arb-alloy 6862abe / arb-revm d3b2517
- harden CI to match arb-alloy/arb-revm: pin toolchain 1.95 (reth MSRV), -D warnings
- migrate ExecutionResult::gas_used -> tx_gas_used in evm tests (identical pre-EIP-8037)
- drop redundant storage_units > 0 guard (COMMIT_THRESHOLD is positive) in snapshot-import
- allow field_reassign (non_exhaustive revm envs), too_many_arguments, and the unread
  provider field kept for reth's converter interface; sort_by_key; mechanical clippy fixes

+++ crates/arb-reth-evm/src/block/tests.rs
@@ -82,7 +82,7 @@ fn exec_cfg() -> ArbExecCfg {
     ArbExecCfg {
         chain_id: CHAIN_ID,
         spec_id: ArbSpecId::ARBOS_51,
-        block_gas_limit: PARENT_GAS_LIMIT as u64,
+        block_gas_limit: PARENT_GAS_LIMIT,
         disable_priority_fee_check: true,
         disable_balance_check: false,
     }
